requirements.txt

torch==1.5.0
torchvision
pillow
numpy
matplotlib
visdom
sklearn




